+++ src/db/mod.rs
@@ -1,4 +1,3 @@
-pub mod action;
 pub mod action_governance;
 pub mod attestation;
 pub mod chisei;
@@ -33,7 +32,6 @@ mod peer_import_store;
 #[cfg(test)]
 mod portfolio_route_test_cases;
 pub mod postgres;
-mod postgres_action;
 mod postgres_action_effect;
 mod postgres_action_governance;
 mod postgres_action_instance;